Added error handling to controller's method

+++ order-service/src/main/java/com/rainbowforest/orderservice/controller/OrderController.java
@@ -1,9 +1,10 @@
 package com.rainbowforest.orderservice.controller;
 
+import com.rainbowforest.orderservice.domain.Item;
+import com.rainbowforest.orderservice.domain.Order;
+import com.rainbowforest.orderservice.domain.User;
 import com.rainbowforest.orderservice.feignclient.UserClient;
-import com.rainbowforest.orderservice.model.Item;
-import com.rainbowforest.orderservice.model.Order;
-import com.rainbowforest.orderservice.model.User;
+import com.rainbowforest.orderservice.http.header.HeaderGenerator;
 import com.rainbowforest.orderservice.service.CartService;
 import com.rainbowforest.orderservice.service.OrderService;
 import com.rainbowforest.orderservice.utilities.OrderUtilities;
@@ -13,6 +14,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.time.LocalDate;
 import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
 
 @RestController
 public class OrderController {
@@ -26,23 +29,46 @@ public class OrderController {
     @Autowired
     private CartService cartService;
 
+    @Autowired
+    private HeaderGenerator headerGenerator;
+    
     @PostMapping(value = "/order/{userId}")
-    private ResponseEntity saveOrder(@PathVariable("userId") Long userId, @RequestHeader(value = "Cookie") String cartId){
-        Order order = new Order();
+    public ResponseEntity<Order> saveOrder(
+    		@PathVariable("userId") Long userId,
+    		@RequestHeader(value = "Cookie") String cartId,
+    		HttpServletRequest request){
+    	
         List<Item> cart = cartService.getAllItemsFromCart(cartId);
-        User user = userClient.getOneUser(userId);
+        User user = userClient.getUserById(userId);   
+        if(cart != null && user != null) {
+        	Order order = this.createOrder(cart, user);
+        	try{
+                orderService.saveOrder(order);
+                cartService.deleteCart(cartId);
+                return new ResponseEntity<Order>(
+                		order, 
+                		headerGenerator.getHeadersForSuccessPostMethod(request, order.getId()),
+                		HttpStatus.CREATED);
+            }catch (Exception ex){
+                ex.printStackTrace();
+                return new ResponseEntity<Order>(
+                		headerGenerator.getHeadersForError(),
+                		HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        }
+  
+        return new ResponseEntity<Order>(
+        		headerGenerator.getHeadersForError(),
+        		HttpStatus.NOT_FOUND);
+    }
+    
+    private Order createOrder(List<Item> cart, User user) {
+        Order order = new Order();
         order.setItems(cart);
         order.setUser(user);
         order.setTotal(OrderUtilities.countTotalPrice(cart));
         order.setOrderedDate(LocalDate.now());
         order.setStatus("PAYMENT_EXPECTED");
-
-        try{
-            orderService.saveOrder(order);
-            cartService.deleteCart(cartId);
-        }catch (Exception ex){
-            ex.printStackTrace();
-        }
-        return new ResponseEntity(order, HttpStatus.CREATED);
+        return order;
     }
 }
